+++ ligo/haskell/test/Test/Ligo/BaseDAO/Proposal/Vote.hs
@@ -1,9 +1,11 @@
 -- SPDX-FileCopyrightText: 2021 TQ Tezos
 -- SPDX-License-Identifier: LicenseRef-MIT-TQ
 
--- | Contains test on @vote@ entrypoint, shared for Lorentz and LIGO contracts.
-module Ligo.BaseDAO.ShareTest.Proposal.Vote
-  ( module Ligo.BaseDAO.ShareTest.Proposal.Vote
+-- | Contains test on @vote@ entrypoint for the LIGO contract.
+module Test.Ligo.BaseDAO.Proposal.Vote
+  ( voteNonExistingProposal
+  , voteMultiProposals
+  , voteOutdatedProposal
   ) where
 
 import Universum
@@ -11,29 +13,29 @@ import Universum
 import Lorentz hiding ((>>))
 import Lorentz.Test hiding (withSender)
 import Morley.Nettest
+import Util.Named
 
-import Ligo.BaseDAO.ShareTest.Common
-import Ligo.BaseDAO.ShareTest.Proposal.Config
+import Test.Ligo.BaseDAO.Common
+import Test.Ligo.BaseDAO.Proposal.Config
 import Ligo.BaseDAO.Types
 
 {-# ANN module ("HLint: ignore Reduce duplication" :: Text) #-}
 
 voteNonExistingProposal
-  :: forall pm param config caps base m.
-    ( MonadNettest caps base m, ProposalMetadataFromNum pm
-    , ParameterContainsEntrypoints param
-      [ "Propose" :> ProposeParams pm
-      , "Vote" :> [PermitProtected $ VoteParam pm]
-      ]
-    , AllConfigDescsDefined config
+  :: forall caps base m.
+    ( MonadNettest caps base m
     , HasCallStack
     )
-  => IsLorentz -> (ConfigDesc config -> OriginateFn param m) -> m ()
+  => IsLorentz -> (ConfigDesc ConfigL -> OriginateFn ParameterL m) -> m ()
 voteNonExistingProposal _ originateFn = do
   ((owner1, _), (owner2, _), dao, _) <- originateFn testConfig
+  advanceTime (sec 10)
+
+  withSender (AddressResolved owner2) $
+    call dao (Call @"Freeze") (#amount .! 2)
 
   -- Create sample proposal
-  _ <- createSampleProposal 1 owner1 dao
+  _ <- createSampleProposal 1 15 owner1 dao
   let params = NoPermit VoteParam
         { vVoteType = True
         , vVoteAmount = 2
@@ -44,18 +46,25 @@ voteNonExistingProposal _ originateFn = do
     & expectCustomErrorNoArg #pROPOSAL_NOT_EXIST
 
 voteMultiProposals
-  :: forall pm param config caps base m.
-    ( MonadNettest caps base m, ParameterC param pm, ProposalMetadataFromNum pm
-    , AllConfigDescsDefined config
+  :: forall caps base m.
+    ( MonadNettest caps base m
     , HasCallStack
     )
-  => IsLorentz -> (ConfigDesc config -> OriginateFn param m) -> m ()
+  => IsLorentz -> (ConfigDesc ConfigL -> OriginateFn ParameterL m) -> m ()
 voteMultiProposals _ originateFn = do
   ((owner1, _), (owner2, _), dao, _) <- originateFn voteConfig
 
+  advanceTime (sec 120)
+  withSender (AddressResolved owner1) $
+    call dao (Call @"Freeze") (#amount .! 20)
+
+  withSender (AddressResolved owner2) $
+    call dao (Call @"Freeze") (#amount .! 5)
+  advanceTime (sec 120)
+
   -- Create sample proposal
-  key1 <- createSampleProposal 1 owner1 dao -- Don't advance time.
-  key2 <- createSampleProposal 2 owner1 dao
+  key1 <- createSampleProposal 1 0 owner1 dao
+  key2 <- createSampleProposal 2 0 owner1 dao
   let params = fmap NoPermit
         [ VoteParam
             { vVoteType = True
@@ -69,23 +78,26 @@ voteMultiProposals _ originateFn = do
             }
         ]
 
+  advanceTime (sec 120)
   withSender (AddressResolved owner2) $ call dao (Call @"Vote") params
   checkTokenBalance (unfrozenTokenId) dao owner2 95
   checkTokenBalance (frozenTokenId) dao owner2 5
   -- TODO [#31]: check storage if the vote update the proposal properly
 
 voteOutdatedProposal
-  :: forall pm param config caps base m.
-    ( MonadNettest caps base m, ParameterC param pm, ProposalMetadataFromNum pm
-    , AllConfigDescsDefined config
+  :: forall caps base m.
+    ( MonadNettest caps base m
     , HasCallStack
     )
-  => IsLorentz -> (ConfigDesc config -> OriginateFn param m) -> m ()
+  => IsLorentz -> (ConfigDesc ConfigL -> OriginateFn ParameterL m) -> m ()
 voteOutdatedProposal _ originateFn = do
   ((owner1, _), (owner2, _), dao, _) <- originateFn testConfig
+  advanceTime (sec 10)
 
+  withSender (AddressResolved owner2) $
+    call dao (Call @"Freeze") (#amount .! 2)
   -- Create sample proposal
-  key1 <- createSampleProposal 1 owner1 dao
+  key1 <- createSampleProposal 1 10 owner1 dao
 
   let params = NoPermit VoteParam
         { vVoteType = True
@@ -93,6 +105,7 @@ voteOutdatedProposal _ originateFn = do
         , vProposalKey = key1
         }
 
+  advanceTime (sec 10)
   withSender (AddressResolved owner2) $ do
     call dao (Call @"Vote") [params]
     advanceTime (sec 25)
